Move pt_fingerprint to utils and test it
