+++ tests/django_mysql_tests/test_utils.py
@@ -1,7 +1,13 @@
 # -*- coding:utf-8 -*-
+from __future__ import unicode_literals
+
+from unittest import skipUnless
+
 from django.test import TestCase
 
-from django_mysql.utils import WeightedAverageRate
+from django_mysql.utils import (
+    have_program, pt_fingerprint, WeightedAverageRate
+)
 
 
 class WeightedAverageRateTests(TestCase):
@@ -37,3 +43,45 @@ class WeightedAverageRateTests(TestCase):
         self.assertEqual(rate.update(50, 0.5), 50)
         self.assertEqual(rate.update(50, 0.5), 50)
         self.assertEqual(rate.update(50, 0.5), 50)
+
+
+@skipUnless(have_program('pt-fingerprint'),
+            "pt-fingerprint must be installed")
+class PTFingerprintTests(TestCase):
+
+    def test_basic(self):
+        self.assertEqual(pt_fingerprint('SELECT 5'), 'select ?')
+        self.assertEqual(pt_fingerprint('SELECT 5;'), 'select ?')
+
+    def test_long(self):
+        query = """
+            SELECT
+                CONCAT(customer.last_name, ', ', customer.first_name)
+                    AS customer,
+                address.phone,
+                film.title
+            FROM rental
+                INNER JOIN customer
+                    ON rental.customer_id = customer.customer_id
+                INNER JOIN address
+                    ON customer.address_id = address.address_id
+                INNER JOIN inventory
+                    ON rental.inventory_id = inventory.inventory_id
+                INNER JOIN film
+                    ON inventory.film_id = film.film_id
+            WHERE
+                rental.return_date IS NULL AND
+                rental_date + INTERVAL film.rental_duration DAY <
+                    CURRENT_DATE()
+            LIMIT 5"""
+        self.assertEqual(
+            pt_fingerprint(query),
+            "select concat(customer.last_name, ?, customer.first_name) as "
+            "customer, address.phone, film.title from rental inner join "
+            "customer on rental.customer_id = customer.customer_id inner join "
+            "address on customer.address_id = address.address_id inner join "
+            "inventory on rental.inventory_id = inventory.inventory_id inner "
+            "join film on inventory.film_id = film.film_id where "
+            "rental.return_date is ? and rental_date ? interval "
+            "film.rental_duration day < current_date() limit ?"
+        )
